https://uci-projects.bitrix24.ru/workgroups/group/16/tasks/task/view/2284/?any=group%2F16%2Ftasks%2Ftask%2Fview%2F2284%2F

+++ back/HR-portal-api/Dal/Repositories/BaseRepository/IBaseRepository.cs
@@ -6,11 +6,11 @@ public interface IBaseRepository<T>
 
     Task<T?> FindWithPredicateAsync(Func<T, bool> predicate);
     
-    Task<T?> FindAsync(int id);
+    Task<T?> FindAsync(long id);
 
     Task CreateAsync(T item);
 
     Task<T> UpdateAsync(T item);
 
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
